folders moved under features folder
