+++ app/src/main/java/com/duzceders/aicaltracker/features/auth/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.duzceders.aicaltracker;
+package com.duzceders.aicaltracker.features.auth;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.duzceders.aicaltracker.R;
 
 public class SignUpActivity extends AppCompatActivity {
